# docs/07-data-api-events/specs/openapi.platform.yaml
openapi: 3.1.2
info:
  title: Korean Equity RWA PoC Platform API
  version: 1.0.0
  description: 합성 데이터만 사용하는 7단계 플랫폼 HTTP 계약이다. 상태변경 명령의 202 응답은 업무 완료가 아니다.
servers:
  - url: http://localhost:3000/api/v1
security:
  - syntheticBearer: []
...
paths:
...
  /workflows/{workflowId}:
    get:
      tags: [CommonQueries]
      operationId: getWorkflow
      summary: 업무의 독립 상태축 조회
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      responses:
        '200':
          description: 독립 상태축과 투영 완전성
          content:
            application/json:
              schema:
                $ref: './schemas/domain.schema.json#/$defs/WorkflowView'
        '404': { $ref: '#/components/responses/Error' }
...
  /dividend-conversions:
    post:
      tags: [InvestorCommands]
      operationId: createDividendConversion
      summary: 지급된 배당의 USDC 전환 요청
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [dividendPaymentId, quoteId]
              properties:
                dividendPaymentId: { type: string, format: uuid }
                quoteId: { type: string, format: uuid }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /voting-instructions:
    post:
      tags: [InvestorCommands]
      operationId: submitVotingInstruction
      summary: 의결권 행사 지시 접수
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [meetingId, agendaId, instruction]
              properties:
                meetingId: { type: string, format: uuid }
                agendaId: { type: string, format: uuid }
                instruction: { type: string, enum: [FOR, AGAINST, ABSTAIN] }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
...
  /reconciliations:
    post:
      tags: [InstitutionCommands]
      operationId: startReconciliation
      summary: 두 축 수량 대사 시작
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [asOf, scope]
              properties:
                asOf: { $ref: './schemas/common.schema.json#/$defs/UtcTimestamp' }
                scope: { type: string, enum: [SECURITY, CUSTOMER, SYSTEM] }
                securityId: { $ref: './schemas/common.schema.json#/$defs/SecurityId' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '409': { $ref: '#/components/responses/Error' }
...
  /holds/{holdId}/release-decisions:
    post:
      tags: [InstitutionCommands]
      operationId: decideHoldRelease
      summary: 중지 해제 승인 또는 거절 기록
      parameters:
        - name: holdId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: './schemas/domain.schema.json#/$defs/InstitutionDecision' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '409': { $ref: '#/components/responses/Error' }
...
  /regulatory-reports/{reportId}/submission-results:
    post:
      tags: [InstitutionCommands]
      operationId: recordRegulatoryReportSubmission
      summary: 월별 보고 제출결과 기록
      parameters:
        - name: reportId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [result, sourceMetadata]
              properties:
                result: { type: string, enum: [ACCEPTED, REJECTED, CORRECTION_REQUIRED] }
                sourceMetadata: { $ref: './schemas/common.schema.json#/$defs/SourceMetadata' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '409': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    syntheticBearer:
      type: http
      scheme: bearer
      bearerFormat: Synthetic-JWT
    adapterSignature:
      type: apiKey
      in: header
      name: X-Adapter-Key-Id
      description: 본문 안의 Ed25519 signature와 함께 검증한다.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128, pattern: '^[A-Za-z0-9._:-]+$' }
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: true
      schema: { type: string, format: uuid }
...
    WorkflowId:
      name: workflowId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Accepted:
      description: 비동기 업무 접수
      content:
        application/json:
          schema: { $ref: './schemas/common.schema.json#/$defs/AcceptedCommand' }
    Error:
      description: 안정적인 기계 오류와 한국어 다음 행동
      content:
        application/json:
          schema: { $ref: './schemas/common.schema.json#/$defs/ErrorResponse' }
...